refactor: 온보딩 모달 부분 컴포넌트로 분리
