+++ src/components/ui/Modal.tsx
@@ -1,3 +1,5 @@
+"use client";
+
 import clsx from "clsx";
 import { useEffect, useRef } from "react";
 
@@ -9,9 +11,13 @@ type Props = {
 };
 
 const ModalItem = ({ isOpen, onClose, children, mode = "default" }: Props) => {
-  const bodyRef = useRef(document.body);
+  // 서버에서 브라우저 객체 undefined 방지
+  const bodyRef = useRef<HTMLElement | null>(null);
 
   useEffect(() => {
+    // 클라이언트에서 초기화
+    bodyRef.current = document.body;
+
     if (isOpen && bodyRef.current) {
       bodyRef.current.style.overflow = "hidden";
     } else if (bodyRef.current) {
